jobs/download_pdf/requirements.txt
google-cloud-storage>=2.0.0,<3.0.0
google-cloud-secret-manager>=2.0.0,<3.0.0
google-auth>=2.0.0,<3.0.0
requests>=2.25.0,<3.0.0
tenacity>=8.0.0,<9.0.0